printing copied boxschedule
